pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prl-hgf"
version = "0.1.0"
description = "HGF-based analysis pipeline for PRL pick_best_cue task with psilocybin intervention"
requires-python = ">=3.10"
authors = [{name = "Adam Manoogian"}]
dependencies = [
    "pyhgf>=0.2.8,<0.3",
    "jax>=0.4.26,<0.4.32",
    "jaxlib>=0.4.26,<0.4.32",
    "numpy>=2.0.0,<3.0",
    "pymc>=5.25.1",
    "pandas>=2.0",
    "scipy>=1.10",
    "matplotlib>=3.4.0",
    "seaborn>=0.13.2",
    "arviz>=0.21.0",
    "ipywidgets>=8.0",
    "ipympl>=0.9.0",
    "pyyaml>=6.0",
    "groupBMC>=1.0",
    "bambi>=0.13.0",
    "pingouin>=0.5.5",
    "pyarrow>=14.0",
    "ptitprince>=0.3.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88
extend-exclude = ["data", "output", "figures", ".planning"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "FA"]
ignore = ["E402", "E501", "E741", "B023", "SIM108"]

[tool.ruff.lint.isort]
known-first-party = ["prl_hgf", "config"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]
"validation/**/*.py" = ["D"]
"scripts/**/*.py" = ["D103"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "numpy.*", "pandas.*", "scipy.*", "matplotlib.*", "seaborn.*",
    "jax.*", "jaxlib.*", "pymc.*", "arviz.*", "pyhgf.*",
    "ipywidgets.*", "bambi.*", "pingouin.*", "ptitprince.*",
    "pyarrow.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests", "validation"]
pythonpath = ["src", "."]
addopts = "-v --tb=short --strict-markers --durations=10"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "requires_pymc: tests that require PyMC installation",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
